refactored dialogue editor into a new module fxgl-tools
