+++ fxgl-samples/pom.xml
@@ -27,18 +27,6 @@
             <artifactId>fxgl-trade</artifactId>
             <version>${fxgl.version}</version>
         </dependency>
-
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.9.7</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.fasterxml.jackson.module</groupId>
-            <artifactId>jackson-module-kotlin</artifactId>
-            <version>2.9.7</version>
-        </dependency>
     </dependencies>
 
     <build>
