+++ fxgl-samples/src/main/java/module-info.java
@@ -5,8 +5,4 @@ open module samples.main {
     // these are only here while being developed
     requires com.almasb.fxgl.trade;
     requires com.almasb.fxgl.cutscene;
-
-    requires com.fasterxml.jackson.databind;
-    requires com.fasterxml.jackson.module.kotlin;
-    requires jackson.annotations;
 }
