+++ fxgl-tools/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+/**
+ * @author Almas Baimagambetov (almaslvl@gmail.com)
+ */
+open module com.almasb.fxgl.tools {
+    requires kotlin.stdlib;
+    requires com.almasb.fxgl.all;
+
+    requires com.fasterxml.jackson.databind;
+    requires com.fasterxml.jackson.module.kotlin;
+    requires jackson.annotations;
+}
