+++ fxgl-tools/src/main/resources/assets/ui/css/dialogue_editor.css
